flask_client/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
grpcio==1.66.1
grpcio-tools==1.66.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
pandas==2.2.2
protobuf==5.28.1
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==74.1.2
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.4
